+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-06-16 10:23:53
-// Commit: test(i18n): add unit tests for i18n business logic
+// Updated at: 2026-06-16 16:44:14
+// Commit: feat(i18n): add bi-directional event bus handler for i18n
 
 class i18nComponentState {
-  final String commitRef = '2014';
-  final String lastMessage = 'test(i18n): add unit tests for i18n business logic';
-  final DateTime timestamp = DateTime.parse('2026-06-16T10:23:53+05:00');
+  final String commitRef = '2033';
+  final String lastMessage = 'feat(i18n): add bi-directional event bus handler for i18n';
+  final DateTime timestamp = DateTime.parse('2026-06-16T16:44:14+05:00');
 }