+++ CHANGELOG.md
@@ -2034,3 +2034,4 @@
 - [2026-06-16 15:04] perf(test): apply RepaintBoundary on animated test widgets
 - [2026-06-16 15:33] refactor(analytics): optimize data flow between domain and data layers
 - [2026-06-16 16:39] fix(navigation): prevent double tap duplicate action in navigation
+- [2026-06-16 16:44] feat(i18n): add bi-directional event bus handler for i18n